feat:-re: Added UI for index.html
Added CSS for <a> element with #search-link-button ID to style it as a button.
Ensured button is centered using a .button-container.
Included hover and active state effects for improved user interaction.
Added names for Flask to scan and PascalCase conventions.
Added image container caption boxes to display descriptions for images.